- Add lookupswitch instruction to enable fast switching with sparse matches - Refactor switch code generation to improve code reuse

+++ model/src/main/java/org/metavm/entity/natives/EnumNative.java
@@ -2,7 +2,7 @@ package org.metavm.entity.natives;
 
 import org.metavm.entity.StdField;
 import org.metavm.object.instance.core.ClassInstance;
-import org.metavm.object.instance.core.LongValue;
+import org.metavm.object.instance.core.IntValue;
 import org.metavm.object.instance.core.StringValue;
 
 public class EnumNative extends NativeBase {
@@ -17,8 +17,8 @@ public class EnumNative extends NativeBase {
         return (StringValue) instance.getField(StdField.enumName.get());
     }
 
-    public LongValue ordinal(CallContext callContext) {
-        return (LongValue) instance.getField(StdField.enumOrdinal.get());
+    public IntValue ordinal(CallContext callContext) {
+        return (IntValue) instance.getField(StdField.enumOrdinal.get());
     }
 
 }
